Moved the mandate.js and mandate.test.js from "src" to "src/mandate" folder
